ssa: makeSlice both len/cap be non-nil
