+++ internal/runtime/goarch/goarch.go
@@ -1,0 +1,3 @@
+package goarch
+
+const PtrSize = 4 << (^uintptr(0) >> 63)
